Exclude 1995 pardons from calculations.
The impact is negligible in stuff like medians, we have fact-checked it, but this is the right approach. Alternatively we could have remove the few 1995 pardons we have from the CSV data file, but we may introduce them next year by mistake. The reason we have 1995 pardons is that we started parsing the BOE from 1st of January 1996, so some pardons were from the previous year.